requirements.txt
# vlm-pid-analysis/requirements.txt

torch
transformers>=4.35.0
accelerate>=0.20.0
bitsandbytes>=0.41.0  
datasets>=2.14.0
huggingface-hub
pillow>=9.0.0
numpy
scipy>=1.10.0 
scikit-learn>=1.3.0
pandas
pyyaml>=6.0
tqdm>=4.65.0